Moved some functions to commonRequest.ts
